fix(agents): keep prepared model completion internal
