+++ src/agents/prepared-model-runtime.configured.ts
@@ -201,41 +201,6 @@ export function prepareConfiguredRuntimeModels(params: {
   return prepared;
 }
 
-export function completeConfiguredRuntimeModels(params: {
-  configuredModelRefs: readonly ConfiguredModelRef[];
-  configuredRuntimeModels: readonly PreparedConfiguredRuntimeModel[];
-  resolveDynamicModel: (lookup: {
-    provider: string;
-    modelId: string;
-  }) => ProviderRuntimeModel | undefined;
-}): PreparedConfiguredRuntimeModel[] {
-  const existing = new Map(
-    params.configuredRuntimeModels.map((configured) => [
-      buildModelCatalogMergeKey(configured.provider, configured.modelId),
-      configured,
-    ]),
-  );
-  const completed: PreparedConfiguredRuntimeModel[] = [];
-  const seen = new Set<string>();
-  for (const { value } of params.configuredModelRefs) {
-    const parsed = parseModelCatalogRef(value);
-    if (!parsed) {
-      continue;
-    }
-    const key = buildModelCatalogMergeKey(parsed.provider, parsed.modelId);
-    if (seen.has(key)) {
-      continue;
-    }
-    seen.add(key);
-    const prepared = existing.get(key);
-    const model = prepared?.model ?? params.resolveDynamicModel(parsed);
-    if (model) {
-      completed.push({ provider: parsed.provider, modelId: parsed.modelId, model });
-    }
-  }
-  return completed;
-}
-
 function findPreparedProviderStaticCatalogModel(params: {
   prepared: PreparedProviderStaticCatalog | undefined;
   metadataSnapshot: PluginMetadataSnapshot;
